pyproject.toml
[tool.poetry]
name = "xqb"
version = "0.1.0"
description = ""
authors = ["adder46 <dedmauz69@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
django-taggit = "^1.3.0"
django-bootstrap4 = "^2.3.1"
django-js-reverse = "^0.9.1"
django-registration = "^3.1.1"
martor = "^1.5.8"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
